Added User UID Binding to Product Records
Updated the add product feature so that when a signed-in "student" type account adds a new product through the AddProducts page, their Firebase User UID is included in the Foreign Key column in the products table. This will be used to determine who is selling a certain product.

Also updated the ProtectedRoutes component to prevent users from being instantly routed to the sign-in page due to the brief moment when the authentication system has not yet recognized the user as signed in. This is done by using Context API.