Back-End Implementation of Product Details Page
Made it so that each individual product containers can be clicked on and further view the information of the product in more detail such as description, reviews, etc.

Also made it so that student users can view a list of their currently listed products.

Also tidied up some of the existing design and deleted some commented code/text.