Added Cart and Address Functionality
Implemented user specific carts which some constraints. Also implemented the back-end part of the user shipping address and added the ability to change a user's shipping address.

Apart from that, tweaked some design to make it tidier and more compatible with the back end. Also removed some elements that are not going to be used.

Also added some toast notifications to indicate successful actions or errors.